ph2rand: Design of randomized comparative phase II oncology trials with a Bernoulli distributed primary outcome

ph2rand provides functions to assist with the design of randomized comparative phase II oncology trials that assume their primary outcome variable is Bernoulli distributed. Specifically, support is provided to (a) perform a sample size calculation when using one of several published designs, (b) evaluate the operating characteristics of a given design (both analytically and via simulation), and (c) produce informative plots.

Getting started

You can install the latest development version of ph2rand from Github with:

devtools::install_github("mjg211/ph2rand")

An introductory example of how to make use of the package's core functionality can be found here. For further help, please see the package vignettes or email michael.grayling@newcastle.ac.uk.

Details

Currently, the following functions are available (exported)

des_one_stage: Determine a one-stage randomized comparative phase II oncology trial design, assuming a Bernoulli primary outcome variable.

des_two_stage: Determine a two-stage randomized comparative phase II oncology trial design, assuming a Bernoulli primary outcome variable.

opchar: Evaluate the operating characteristics (analytically) of a randomized comparative phase II oncology trial design that assumes a Bernoulli primary outcome variable.

plot.ph2rand_des: Plot the operating characteristics of a randomized comparative phase II oncology trial design that assumes a Bernoulli primary outcome variable.

plot.ph2rand_pmf: Plot the probability mass function of a randomized comparative phase II oncology trial design that assumes a Bernoulli primary outcome variable.

plot.ph2rand_terminal: Plot the terminal points of a randomized comparative phase II oncology trial design that assumes a Bernoulli primary outcome variable.

pmf: Find the probability mass function of a randomized comparative phase II oncology trial design that assumes a Bernoulli primary outcome variable.

sim: Evaluate the operating characteristics (via simulation) of a randomized comparative phase II oncology trial design that assumes a Bernoulli primary outcome variable.

summary.ph2rand_des: Display a summary of a randomized comparative phase II oncology trial design that assumes a Bernoulli primary outcome variable.

terminal: Find the terminal points of a randomized comparative phase II oncology trial design that assumes a Bernoulli primary outcome variable.

References

Jung SH (2008) Randomized phase II trials with a prospective control. Stat Med 27(4):568--83. DOI: 10.1002/sim.2961. PMID: 17573688.

Jung SH, Sargent DJ (2014) Randomized phase II clinical trials. J Biopharm Stat 24(4):802--16. DOI: 10.1080/10543406.2014.901343. PMID: 24697589.

Kepner JL (2010) On group sequential designs comparing two binomial proportions. J Biopharm Stat 20(1):145--59. DOI: 10.1080/10543400903280621. PMID: 20077254.

Litwin S, Basickes S, Ross EA (2017) Two-sample binary phase 2 trials with low type I error and low sample size. Stat Med 36(9):1383--94. DOI: 10.1002/sim.7226. PMID: 28118686.

Shan G, Ma C, Hutson AD, Wilding GE (2013) Randomized two-stage phase II clinical trial designs based on Barnard's exact test. J Biopharm Stat 23(5):1081--90. DOI: 10.1080/10543406.2013.813525. PMID: 23957517.